backend/requirements.txt
flask
flask-cors
twilio
ngrok
python-dotenv